+++ src/main/java/com/koshelev/spring/web/dto/ProductDto.java
@@ -12,6 +12,6 @@ public class ProductDto {
 
     private Long id;
     private String title;
-    private Double cost;
+    private Integer cost;
 
 }
